requirements.txt
# ETL Process Dependencies

# Core data processing
pandas>=1.5.0
numpy>=1.20.0

# Configuration and file handling
PyYAML>=6.0
openpyxl>=3.0.0

# Optional: for advanced features
# requests>=2.28.0  # For API data extraction
# sqlalchemy>=1.4.0  # For database operations
# psycopg2-binary>=2.9.0  # For PostgreSQL
# pymongo>=4.0.0  # For MongoDB